Added sex of the cat. App name change.
Added the ability for the cat to be either male or female, and added a
UI component to display which one the cat is. Right now, upon running
the program, the cat's sex will be randomly decided.
Changed the name of the app to CatNip (working title).
Changed the icon of the app to be the cat face pixel art sprite.