headTemplate and moodPercentageVariable
added two things: a headerTemplate text document in the top directory of
the project so that we can nail a complete header and get it sorted out
early.
Second, added a double to all the cat's attributes other than mood that
will act as a percentage so that when the cat is fed, given water, or
groomed(if hygiene is implemented), the cat's mood can go up base don a
percent of the incrementAmount value of the attribute that went up. It
could also go down by that percentage is the cat gets more hungry,
thirsty, or unhygienic.
The moodPercentageVariable is a little vague, but I can explain it in
detail if need be, hopefully this description is enough for now.