Fixed the Meter progress being lost.
Fixed the issue where the Meters' progress was lost whenever the onCreate method was called.
Pretty much all of the code used in the meters was restructured, generalized, and moved into the Meter class. This means that a lot of code duplication is no longer present, and changes will be easier. Now, you don't need to implement the methods in the separate meter classes, but you can instead override the methods if necessary.
Added an Init class that holds globals and performs most of the initialization before the Activity is initialized.
Changed a few minor things to get ready for future code.