Fixed screen dimensions.
Changed all the screen dimensions to be measured in dp and sp instead of pt. This allows for correct scalability.
Added dimensions to dimens.xml.
Added a few low dpi images (just non-scaled-up versions of the original pixel art images). Will add the rest later.
Changed the default mood to Content for testing purposes (it's the longest mood text and I wanted to make sure it didn't run into the progress bar).