Initial saving of data and settings activity implemented
Implemented some framework for basic saving functionality. Currently, if you haven't set the sex of the cat, it will ask you to set it and it will be saved until you clear the data for the cat.
Added a rudimentary, but somewhat functional settings screen. Cat data can be cleared from the settings screen.
Reorganized/renamed a bunch of stuff.
Removed some unneeded things.
Tried to clean up some of the code.