Implemented SharedPreferences for persistent data
Implemented SharedPreferences so that we can have persistent data to store things like settings, cat data, player data, and anything that we want to be persistent across app launches.
Added a setup wizard to ask you what you want to name your cat and what you want the sex of the cat to be if it detects that you don't already have it set up.
Updated just about everything to use the strings.xml instead of hard-coded strings. This will make things easier to change and will help if we decide to localize our app.
Fixed memory leaks.