Petting Implementation and Bug Fix
First, a bug fix: in the AndroidManifest.xml file, the SettingsActivity
had the meta-data - android::value information wrong. It was trying to
look into
com.hogg.catapps.ui.MainActivity/com.hogg.catapps.ui.MainActivity, and
was making my emulator crash. I sincerely hope that it is NOT an
emulator problem, as that would mean I've changed it to be a bug.
Regardless, we should look into it further with Drew's phone.
Secondly, in regards to the petting implementation, I did not use
gestures at all, I used MotionEvents because I found the documentation
much more clear for it, and it's got a lot of flexibility.
The way petting is implemented now is thus: In MainActivity.onCreate() I
create a RelativeLayout called petLayout. This gets the ID for the
overall relativelayout in our layout(I used the same ID for the topmost
relative layouts in the layout-land and layout so that the program can
get it dynamically without any logic checking).
Once the RelativeLayout has been identified and stored, I set it's
OnTouchListener to one of my own design, PetListener.
**Something of importance to note here is that since it is simply
setting the OnTouchListener of the relative layout, it does NOT need to
be done in onResume() and then destroyed in onPause(), as once it is
instantiated it, that relative layout's OnTouchListener will not
change.**
Every time an ACTION_DOWN is caught, it is in a new motion. It stores
the X and Y values of the action and waits for the ACTION_UP. While
waiting on the ACTION_UP, if any ACTION_MOVE occurs, it is tested to see
if it is within the confines of the rectangle given by the ImageView
imageCat. The boolean true or false is then stored in a List, for
calculations later.
Once the ACTION_UP is seen, the following calculations occur:
1) If the ACTION_DOWN and ACTION_UP were both within the confines of the
rectangle, the onCatButtonClick method is called and the text Meow!
appears and the hearts go up by 5.
2) The number of times an ACTION_MOVE event is within the confines of
the ImageView imageCat is divided by the number of times an ACTION_MOVE
event occurs, effectively getting the percentage of ACTION_MOVE events
that occur "on" the "cat."
3) If this percentage is greater than or equal to the percentPet, then
it is a pet. Hearts is incremented by 10 and updated.
4) Regardless of whether or not the whole MotionEvent sequence was
classified as a pet or not, the values in all the tracking variables is
cleared for the next pet.

I also put Drew's name in contributors for the PetListener as the
onCatButtonClick method was his design.