State Machine Simulation
This should, in theory, work completely.
See notes in the src\com\hogg\catapps\simulation\Simulation.java file
for more info on how to use this.
It is a self contained system and needs to be handled from the outside
manually to function correctly with other portions of the application.