PetListener encapsulation and extensions
Encapsulated the code for decreasing the hearts due to a bad touch event
such as poking the cat, or petting against the grain.
Added functionality to determine if the pet is against the grain of the
cats fur. Is only taken into account if the difficulty is set to
Hardcore.